Diagnosed and fixed an issue with not reliably completing command sequence at HP station.
The elevator is stopped inconsistently by the bottom magnet limit switch. Combined with slack in the chain, this can cause the elevator to stop with the primary motor still above 1 rotation away from its zero, preventing the sequence from advancing to run intake, then zero, etc. Increasing the elevator tolerance to 2 seems to prevent the issue, and because of the "hold" command behavior, the motor doesn't actually stop fighting to reach a more exact position after the command completes.